backend/requirements.txt
fastapi
uvicorn
yfinance
pandas
ta
requests
python-dotenv
